epic_1_include_mvp
This release brings the basic API with the essential resources required to perform coalescing operations. The docstrings are still not fully covered. The __repr__ and __str__ magic methods are also not fully covered. The API.md file and the /examples directory won't be filled at this release. And lastly, the unit tests haven't still been elaborated.

- Features:
    - Special Objects:
        1. VoidTypes
    - Special Types
        1. UndefinedType
    - Special Constants
        1. UNDEFINED - alias: Undefined
        2. VOID - alias: Void
    - Special Agents
        1. VOIDSAFE - alias: VoidSafe
        2. VALUE - alias: value
    - Special Operators
        1. IFNONE - alias: ifnone
        2. IFNDEF - alias: ifndef
        3. IFVOID - alias: ifvoid

- #1: Include Project with MVP
- Gustavo Cardoso Ribeiro
- 2024-03-10